requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
pipreqs==0.4.11
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
yarg==0.1.9
zipp==3.7.0
